setezor/requirements.txt
aiosqlite==0.19.0
fastapi==0.114.2
uvicorn==0.30.6
SQLAlchemy==2.0.34
aiojobs==1.3.0
pydantic-extra-types==2.9.0
pydantic==2.9.1
dnspython==2.6.1
cpeguess @ git+https://github.com/lmsecure/CPEGuess@38506e0f96d57b6d0cbc6a3ab4ed83a22ec2a99c
passlib==1.7.4
sqlmodel==0.0.22
PyJWT==2.10.0
Jinja2==3.1.4
aiofiles==24.1.0
aiohttp==3.11.7
pyOpenSSL==24.2.1
python-multipart==0.0.17
python-dotenv==1.0.1
orjson==3.10.12
aiodns==3.2.0
xmltodict==0.13.0
pyroute2==0.7.12
aiohttp-jinja2==1.6
websockets==14.1
scapy==2.5.0
pycryptodome==3.21.0
playwright==1.47.0
